chore: add mutator-common crate
This crate contains the common tools shared by `move-test-spec` and
`move-mutation-test` tools.

In this commit, we only add:
- report module

To be added in the future:
- benchmark module
