+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "move-mutation-test",
     "move-mutator",
     "move-spec-test",
+    "mutator-common",
 ]
 
 # All workspace members should inherit these keys
@@ -45,6 +46,7 @@ move-symbol-pool = { git = "https://github.com/aptos-labs/aptos-core.git", branc
 move-unit-test = { git = "https://github.com/aptos-labs/aptos-core.git", branch = "main" }
 move-vm-runtime = { git = "https://github.com/aptos-labs/aptos-core.git", branch = "main" }
 num = "0.4"
+mutator-common = { path = "mutator-common" }
 num-traits = "0.2"
 pretty_env_logger = "0.5"
 rand = "0.8"
