+++ move-mutator/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 description = "Move mutation tool"
 authors.workspace = true
 edition.workspace = true
+homepage.workspace = true
 license.workspace = true
 publish.workspace = true
 repository.workspace = true
