+++ mutator-common/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "mutator-common"
+version = "0.1.0"
+description = "Common set of tools for mutator applications"
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+publish.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+tabled = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
