+++ mutator-common/src/lib.rs
@@ -1,0 +1,6 @@
+// Copyright © Eiger
+// Copyright © Aptos Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+/// A module for generating concise, valuable reports.
+pub mod report;
